Removed usage of --no-sound-null-safety flag
