+++ tool/coverage.sh
@@ -11,7 +11,7 @@ set -e
 pub global activate remove_from_coverage
 pub global activate coverage
 # Generate coverage report.
-dart --no-sound-null-safety --pause-isolates-on-exit --disable-service-auth-codes --enable-vm-service=3000 test/test_coverage.dart &
+dart --pause-isolates-on-exit --disable-service-auth-codes --enable-vm-service=3000 test/test_coverage.dart &
 dart pub global run coverage:collect_coverage --wait-paused --uri=http://127.0.0.1:3000/ -o coverage.json --resume-isolates
 dart pub global run coverage:format_coverage --lcov -i coverage.json -o lcov.info
